update conditional compilation in tests
