+++ tests/ui/abi3_dict.stderr
@@ -1,4 +1,4 @@
-error: `dict` requires >= python 3.9 with the `abi3` feature
+error: `dict` requires Python >= 3.9 when using the `abi3` feature
  --> tests/ui/abi3_dict.rs:4:11
   |
 4 | #[pyclass(dict)]
